pyproject.toml
[tool.poetry]
package-mode = false
name = "sell-that-sheet-api"
version = "0.1.0"
description = ""
authors = ["wsmcrn <zs1cyranik@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
Django = "^5.0.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
